requirements.txt
asgiref==3.6.0
mysqlclient==2.1.1
cffi==1.15.1
cryptography==39.0.0
Django==4.1.5
djangorestframework==3.14.0
greenlet==2.0.1
numpy==1.24.1
pandas==1.5.2
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.4.3
tzdata==2022.7
